+++ src/tft.h
@@ -92,7 +92,7 @@ typedef struct
                       uint16_t size, uint16_t fgcolor, uint16_t bgcolor);
     void (*draw_string)(char *string, uint16_t x, uint16_t y,
                         uint16_t size, uint16_t fgcolor, uint16_t bgcolor);
-    uint8_t (*draw_number)(int64_t long_num, uint16_t x, uint16_t y, 
+    uint8_t (*draw_number)(int long_num, uint16_t x, uint16_t y, 
                            uint16_t size, uint16_t fgcolor, uint16_t bgcolor);
     void (*draw_char_only)(uint8_t ascii, uint16_t x, uint16_t y, 
                            uint16_t size, uint16_t color);
